backend/requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.48
alembic==1.11.1
pydantic==2.10.5
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.0
email-validator==2.0.0
psycopg2-binary==2.9.10
python-multipart==0.0.22
